requirements.txt
numpy
opencv-python
matplotlib
torch
pyyaml
pytorch-lightning